+++ packages/lumx-core/src/js/types/JSXElement.ts
@@ -1,5 +1,13 @@
 import type React from 'react';
 
-export type JSXElement = boolean | number | string | React.JSX.Element | Iterable<JSXElement> | undefined | null;
+export type JSXElement =
+    | boolean
+    | number
+    | string
+    | React.JSX.Element
+    | React.ReactNode
+    | Iterable<JSXElement>
+    | undefined
+    | null;
 
 export type ElementType = React.ElementType;