+++ packages/lumx-vue/src/utils/ResetTheme.ts
@@ -1,0 +1,9 @@
+import { defineComponent, provide } from 'vue';
+
+export const ResetTheme = defineComponent({
+    name: 'ResetTheme',
+    setup(_, { slots }) {
+        provide('theme', undefined);
+        return () => slots.default?.();
+    },
+});